Support for RFQ-T indicative quotes
Renamed method getSwapQuoteAsync() to _getSwapQuoteAsync(), and created
a new getSwapQuoteAsync() to delegeate to it, all for the purposes of
also calling into _getSwapQuoteAsync() of the newly introduced method
getSwapPriceAsync().

Unfortunately, the diff looks noisy because the old public method was
changed to be private, forcing it to move down below publics in the
method order.  This resulted resulting in my diff viewer showing that
two previously existing methods (getSwapPricesAsync() [notice that one's
plural, not to be confused with the aforementioned singular version] and
getTokenPricesAsync()) moved UP in the diff.

Those two existing methods have not changed, and it's easier to review
this commit if you understand the pattern they contribute to the overall
diff.