Log orders underlying indicative quotes
Log the orders that are used to construct indicative quotes.  These are
"dummy" orders, but for now we need to see them in the logs in order to
verify that indicative quotes are being served, since there is no
separate source for RFQ-T (they are lumped in with the '0x' source.)

Long term, this log entry may prove useful if it isn't too noisy.  If it
is, we can remove it after manual testing.